requirements.txt
Jinja2==2.11.2
PyYAML==6.0
requests==2.28.1
schedule==1.1.0
matplotlib==3.5.3
